Move nodejs-utils under sonar-javascript-plugin
