+++ pom.xml
@@ -32,7 +32,6 @@
     <module>javascript-checks</module>
     <module>sonar-javascript-plugin</module>
     <module>eslint-bridge</module>
-    <module>nodejs-utils</module>
   </modules>
 
   <scm>
@@ -86,11 +85,6 @@
       <dependency>
         <groupId>${project.groupId}</groupId>
         <artifactId>javascript-checks</artifactId>
-        <version>${project.version}</version>
-      </dependency>
-      <dependency>
-        <groupId>${project.groupId}</groupId>
-        <artifactId>nodejs-utils</artifactId>
         <version>${project.version}</version>
       </dependency>
       <dependency>
