+++ sonar-javascript-plugin/src/main/java/org/sonar/plugins/javascript/nodejs/NodeCommandException.java
@@ -17,11 +17,15 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonarsource.nodejs;
+package org.sonar.plugins.javascript.nodejs;
 
-@FunctionalInterface
-public interface BundlePathResolver {
+public class NodeCommandException extends RuntimeException {
 
-  String resolve(String relativePath);
+  public NodeCommandException(String message) {
+    super(message);
+  }
 
+  public NodeCommandException(String message, Throwable cause) {
+    super(message, cause);
+  }
 }
