+++ sonar-javascript-plugin/src/main/java/org/sonar/plugins/javascript/nodejs/package-info.java
@@ -18,4 +18,4 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 @javax.annotation.ParametersAreNonnullByDefault
-package org.sonarsource.nodejs;
+package org.sonar.plugins.javascript.nodejs;
