+++ sonar-javascript-plugin/src/test/resources/files/bundle.txt
