+++ sonar-javascript-plugin/src/test/resources/files/error.js
