+++ sonar-javascript-plugin/src/test/resources/files/script.js
