+++ sonar-javascript-plugin/src/test/resources/run-node/run-node
